Fix: Move JavaScript from non-existent extra_scripts block to extra_head
The base.html template doesn't have an 'extra_scripts' block - it has
'extra_head' and 'extra_body'. The JavaScript code for loading integrations
was in a non-existent block and therefore wasn't being rendered.

Moved the script to the 'extra_head' block so it's properly included
in the page template. This should fix the integration selector not working
since the JavaScript wasn't even loading.

This was the root cause of the integration dropdown not responding to
project selection.